refactor: 💡 move modal to assignments
