+++ apps/hems/src/app/app.module.ts
@@ -14,7 +14,7 @@ import { CheckinComponent } from './pages/checkin/checkin.component';
 import { DashboardComponent } from './pages/dashboard/dashboard.component';
 import { PageNotFoundComponent } from './pages/page-not-found/page-not-found.component';
 import { AssignmentsPageComponent } from './pages/assignments/assignments-page.component';
-import { CreateAssignmentModalComponent } from './components/modals/createAssignmentModal/create-assignment-modal.component';
+import { CreateAssignmentModalComponent } from './pages/assignments/createAssignmentModal/create-assignment-modal.component';
 
 @NgModule({
   declarations: [
