+++ apps/hems/src/app/pages/assignments/createAssignmentModal/create-assignment-modal.component.html
@@ -11,7 +11,6 @@
         matInput
         placeholder="Room nr"
         maxlength="10"
-        required
       />
       <mat-error
         *ngIf="
@@ -32,7 +31,6 @@
         matInput
         placeholder="Type of task"
         maxlength="100"
-        required
       />
       <mat-error
         *ngIf="
@@ -53,7 +51,6 @@
         name="comments"
         placeholder="Describe the task"
         maxlength="1000"
-        required
       ></textarea>
       <mat-error
         *ngIf="
@@ -74,7 +71,6 @@
         matInput
         placeholder="Received by"
         maxlength="100"
-        required
       />
       <mat-error
         *ngIf="
@@ -95,7 +91,6 @@
         matInput
         placeholder="Performed by"
         maxlength="100"
-        required
       />
       <mat-error
         *ngIf="
@@ -114,7 +109,6 @@
         formControlName="receivedAt"
         name="receivedAt"
         matInput
-        required
       />
       <mat-error
         *ngIf="
