+++ apps/hems/src/app/pages/assignments/createAssignmentModal/create-assignment-modal.component.scss
